Change how iot is loaded
