+++ projects/social-host/src/app/controls/iot/iot.component.ts
@@ -53,31 +53,5 @@ export class IoTComponent implements OnInit {
   }
 
   //  Helpers
-  protected async handleStateChange(): Promise<void> {
-    this.loadScripts();
-
-    this.loadStyles();
-  }
-
-  protected loadScripts() {
-    for (let script of this.IoTConfig.Scripts) {
-      let node = document.createElement('script'); // creates the script tag
-      node.src = script; // sets the source (insert url in between quotes)
-      node.type = 'text/javascript'; // set the script type
-      node.async = true; // makes script run asynchronously
-      node.charset = 'utf-8';
-      // append to head of document
-      document.getElementsByTagName('head')[0].appendChild(node);
-    }
-  }
-
-  protected loadStyles() {
-    for (let style of this.IoTConfig.Styles) {
-      let node = document.createElement('link'); // creates the script tag
-      node.href = style; // sets the source (insert url in between quotes)
-      node.type = 'text/css'; // set the script type
-      // append to head of document
-      document.getElementsByTagName('head')[0].appendChild(node);
-    }
-  }
+  protected async handleStateChange(): Promise<void> {}
 }
